+++ pcr2/src/main/kotlin/edu/monash/humanise/smartcity/pcr2/PayloadRepository.kt
@@ -2,4 +2,4 @@ package edu.monash.humanise.smartcity.pcr2
 
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface PayloadRepository : JpaRepository<Payload, Int>
+interface PayloadRepository : JpaRepository<Payload, Long>
